Change CashOnDelivery value type

+++ lib/core/model/tracking_service/ups/ups_parser.dart
@@ -874,7 +874,7 @@ class _CashOnDelivery {
   const _CashOnDelivery({required this.value, required this.currencyCode});
 
   CashOnDelivery toCashOnDelivery() =>
-      CashOnDelivery(int.parse(value), currencyCode);
+      CashOnDelivery(int.parse(value).toDouble(), currencyCode);
 }
 
 abstract class _ShipmentAddressTypeCode {
